+++ R/defaultPlotConfig.R
@@ -43,6 +43,7 @@
 #' @param label_max_overlaps Maximum number of overlapping labels.
 #' @param label_nudge_x Horizontal nudge for label text.
 #' @param label_nudge_y Vertical nudge for label text.
+#' @param label_nudge_y_flip TRUE. Whether to flip the nudge y value to be negative. The plot is reversed vertically, so this is needed to nudge labels up instead of down.
 #' @param label_segment_color Segment color for label connectors.
 #' @param label_text_angle Text angle for labels.
 #' @param label_text_size Font size for labels.
@@ -189,33 +190,17 @@ getDefaultPlotConfig <- function(function_name = "getDefaultPlotConfig",
                                  value_rounding_digits = 5,
                                  # --- SEX ------------------------------------------------------------
                                  code_male = 1,
-                                 # ---- Filtering and Computation ----
-                                 filter_n_pairs = 500,
-                                 filter_degree_min = 0,
-                                 filter_degree_max = 7,
-                                 drop_classic_kin = FALSE,
-                                 drop_non_classic_sibs = TRUE,
-                                 use_only_classic_kin = TRUE,
-                                 use_relative_degree = TRUE,
-                                 group_by_kin = TRUE,
-                                 # ----Kinbin Settings ----
-                                 match_threshold_percent = 10,
-                                 max_degree_levels = 12,
-                                 grouping_column = "mtdna_factor",
-                                 # ---- Annotation Settings ----
-                                 annotate_include = TRUE,
-                                 annotate_x_shift = -0.1,
-                                 annotate_y_shift = 0.005,
                                  # ---- Label Aesthetics ----
                                  label_include = TRUE,
                                  label_column = "personID",
                                  label_method = "ggrepel",
                                  label_max_overlaps = 15,
                                  label_nudge_x = 0,
-                                 label_nudge_y = -0.10,
+                                 label_nudge_y = 0.10,
+                                 label_nudge_y_flip = TRUE, # flip the nudge y value to be negative
                                  label_segment_color = NA,
                                  label_text_angle = 0,
-                                 label_text_size = 2,
+                                 label_text_size = 3,
                                  label_text_color = "black",
                                  label_text_family = "sans",
                                  # --- POINT / OUTLINE AESTHETICS ---------------------------------------
@@ -243,7 +228,7 @@ getDefaultPlotConfig <- function(function_name = "getDefaultPlotConfig",
                                  ped_align = TRUE,
                                  ped_width = 15,
                                  # ---- Segment Drawing Options ----
-                                 segment_linewidth = 0.5,
+                                 segment_linewidth = .5,
                                  segment_linetype = 1,
                                  segment_lineend = "round",
                                  segment_linejoin = "round",
@@ -253,11 +238,11 @@ getDefaultPlotConfig <- function(function_name = "getDefaultPlotConfig",
                                  segment_sibling_color = segment_default_color,
                                  segment_spouse_color = segment_default_color,
                                  segment_mz_color = segment_default_color,
-                                 segment_mz_linetype = 1,
+                                 segment_mz_linetype = segment_linetype,
                                  segment_mz_alpha = 1,
                                  segment_mz_t = .6,
                                  segment_self_linetype = "dotdash",
-                                 segment_self_linewidth = .25,
+                                 segment_self_linewidth = .5*segment_linewidth,
                                  segment_self_alpha = 0.5,
                                  segment_self_angle = 90,
                                  segment_self_curvature = -0.2,
@@ -325,7 +310,26 @@ getDefaultPlotConfig <- function(function_name = "getDefaultPlotConfig",
                                  ),
                                  focal_fill_labels = c("Low", "Mid", "High"),
                                  # Use first color for affected,
-                                 # ---- Confidence Intervals
+
+                                 # ---- correlation by bin ----
+                                 ## ---- Filtering and Computation ----
+                                 filter_n_pairs = 500,
+                                 filter_degree_min = 0,
+                                 filter_degree_max = 7,
+                                 drop_classic_kin = FALSE,
+                                 drop_non_classic_sibs = TRUE,
+                                 use_only_classic_kin = TRUE,
+                                 use_relative_degree = TRUE,
+                                 group_by_kin = TRUE,
+                                 ## ----Kinbin Settings ----
+                                 match_threshold_percent = 10,
+                                 max_degree_levels = 12,
+                                 grouping_column = "mtdna_factor",
+                                 ## ---- Annotation Settings ----
+                                 annotate_include = TRUE,
+                                 annotate_x_shift = -0.1,
+                                 annotate_y_shift = 0.005,
+                                 ## ---- Confidence Intervals
                                  ci_include = TRUE,
                                  ci_ribbon_alpha = .3,
                                  # ---- tile settings ----
@@ -421,6 +425,7 @@ getDefaultPlotConfig <- function(function_name = "getDefaultPlotConfig",
     label_max_overlaps = label_max_overlaps,
     label_nudge_x = label_nudge_x,
     label_nudge_y = label_nudge_y,
+    label_nudge_y_flip = label_nudge_y_flip,
     label_segment_color = label_segment_color,
     label_text_angle = label_text_angle,
     label_text_size = label_text_size,
@@ -575,12 +580,13 @@ getDefaultPlotConfig <- function(function_name = "getDefaultPlotConfig",
   if (lc_function_name %in% c("ggrelatednessmatrix")) {
     #   If the function is ggRelatednessMatrix, we need to adjust the tooltip columns
     core_list$tooltip_columns <- c("ID1", "ID2", "value")
-
+    core_list$label_nudge_y_flip <- FALSE
     #  core_list$tile_color_palette <- c(
     #    core_list$color_palette_low,
     #   core_list$color_palette_mid,
     #   core_list$color_palette_high
     # )
+
     core_list$color_scale_midpoint <- 0.25
     core_list$plot_title <- "Relatedness Matrix"
     core_list$axis_x_label <- "Individual"
@@ -596,7 +602,7 @@ getDefaultPlotConfig <- function(function_name = "getDefaultPlotConfig",
     core_list$return_static <- FALSE
     core_list$return_widget <- FALSE
     core_list$return_interactive <- FALSE
-
+    core_list$label_nudge_y_flip <- FALSE
     #  default_config <- list(
     #    apply_default_scales = TRUE,
     #    apply_default_theme = TRUE,
@@ -634,7 +640,7 @@ getDefaultPlotConfig <- function(function_name = "getDefaultPlotConfig",
   )) {
     core_list$label_method <- "ggrepel"
     core_list$label_column <- personID
-
+    core_list$label_nudge_y_flip <- TRUE
     # core_list$focal_fill_low_color <- core_list$color_palette_low
     # core_list$focal_fill_mid_color <- core_list$color_palette_mid
     # core_list$focal_fill_high_color <- core_list$color_palette_high
@@ -649,12 +655,15 @@ getDefaultPlotConfig <- function(function_name = "getDefaultPlotConfig",
     core_list$tooltip_columns <- c(personID, "sex", status_column)
     core_list$label_method <- "geom_text"
     core_list$label_include <- FALSE # default to FALSE
+    core_list$segment_linewidth <- 0.5 # too think
+    core_list$segment_self_linewidth <- .5*core_list$segment_linewidth
     core_list$tooltip_include <- TRUE
     core_list$return_static <- FALSE
     core_list$return_widget <- TRUE
     core_list$return_interactive <- TRUE
     core_list$segment_self_angle <- -75
     core_list$segment_self_curvature <- -0.15
+
   }
 
   return(core_list)
@@ -697,6 +706,9 @@ buildPlotConfig <- function(default_config,
 
   # -- Merge user config with defaults --
   built_config <- utils::modifyList(default_config, config)
+
+  built_config$label_nudge_y <- ifelse(built_config$label_nudge_y_flip,
+    built_config$label_nudge_y*-1,built_config$label_nudge_y)
 
   if (stringr::str_to_lower(function_name) %in%
     c("ggpedigree", "ggpedigreeinteractive")) {
@@ -746,7 +758,7 @@ buildPlotConfig <- function(default_config,
   } else if (stringr::str_to_lower(function_name) %in%
     c("ggphenotypebydegree", "phenotypebydegree")) {
 
-
+    built_config$label_nudge_y_flip <- FALSE # default to TRUE for ggphenotypebydegree
   }
 
   return(built_config)