+++ tests/testthat/test-defaultPlotConfig.R
@@ -16,7 +16,7 @@ test_that("getDefaultPlotConfig returns expected defaults", {
   config <- getDefaultPlotConfig()
 
   expect_true(is.list(config))
-  expect_equal(length(config), 151) # Check number of default parameters
+  expect_equal(length(config), 154) # Check number of default parameters
 
   expect_equal(config$apply_default_scales, TRUE)
   expect_equal(config$apply_default_theme, TRUE)
@@ -58,6 +58,8 @@ test_that("handles function_name variations", {
   config1 <- getDefaultPlotConfig(function_name = "ggpedigree")
   config2 <- getDefaultPlotConfig(function_name = "ggPedigreeInteractive")
   config3 <- getDefaultPlotConfig(function_name = "ggrelatednessmatrix")
+
+
   expect_true(is.list(config1))
   expect_true(is.list(config2))
   expect_true(is.list(config3))
@@ -66,6 +68,10 @@ test_that("handles function_name variations", {
 
   expect_equal(config1$label_method, "ggrepel")
   expect_equal(config2$label_method, "geom_text")
+  expect_equal(config2$label_nudge_y_flip, TRUE)
+  expect_equal(config3$label_nudge_y_flip, FALSE)
+  expect_equal(config1$label_nudge_y, config2$label_nudge_y)
+  expect_equal(config1$label_nudge_y, config3$label_nudge_y)
   expect_true(config2$return_widget)
   expect_true(config2$return_interactive)
   expect_false(config2$return_static)
@@ -115,11 +121,14 @@ test_that("buildPlotConfig merges valid subset overrides correctly", {
     result$segment_linewidth,
     default_config$segment_linewidth
   )
+  expect_equal(result$label_nudge_y, -0.10)
+  expect_equal(result$label_nudge_y, -1*default_config$label_nudge_y)
 })
 test_that("buildPlotConfig returns a list", {
-  default_config <- getDefaultPlotConfig(function_name = "ggPedigree")
+  default_config <- getDefaultPlotConfig(function_name = "ggrelatednessmatrix")
   custom <- list(point_size = 2)
 
   result <- buildPlotConfig(default_config, custom)
+  expect_equal(result$label_nudge_y, 0.10)
   expect_true(is.list(result))
 })