+++ NEWS.md
@@ -7,6 +7,8 @@
 * Smarter implementation of aliases
 * Updated hex to actually use ggpedigree
 * exposed relation and hints inputs to ggpedigree
+* Exposed 'relations' and 'hints' inputs to ggpedigree
+* Added more intuitive defaults
 
 # ggpedigree 0.8.0
 * Makes ggrepel a suggested package instead of a required one.